Поменял логику в контролерах, по мелочи изменил сервисный слой и репозитории

+++ src/main/java/ru/kata/spring/boot_security/demo/repositories/RoleRepository.java
@@ -2,11 +2,11 @@ package ru.kata.spring.boot_security.demo.repositories;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
-import ru.kata.spring.boot_security.demo.models.Role;
-import ru.kata.spring.boot_security.demo.models.User;
+import ru.kata.spring.boot_security.demo.entities.Role;
 
-import java.util.Optional;
+import java.util.Set;
 
 @Repository
-public interface RoleRepository extends JpaRepository<Role, Integer> {
+public interface RoleRepository extends JpaRepository<Role, Long> {
 }
+
